requirements.txt
# 基础依赖
python >= 3.8
numpy >= 1.21.0
torch >= 1.13.0
torchvision >= 0.14.0
torchaudio >= 0.13.0

# 数据处理
librosa >= 0.9.2
opencv-python >= 4.5.5
dlib >= 19.23.0
pillow >= 9.0.1
soundfile >= 0.10.3.post1
einops >= 0.4.1

# 模型训练与评估
tqdm >= 4.64.0
scikit-learn >= 1.0.2
tensorboard >= 2.11.0
PyYAML >= 6.0

# Web服务
gradio >= 3.16.2
fastapi >= 0.95.0
uvicorn >= 0.21.1
moviepy >= 1.0.3  # 用于视频音频合并

# 可选加速库
cudatoolkit >= 11.6  # 如使用GPU
onnxruntime >= 1.13.1  # 如使用ONNX推理
    